+++ web/cm_plugins/command_link.ts
@@ -1,4 +1,3 @@
-import { commandLinkRegex } from "../../common/markdown_parser/parser.ts";
 import { ClickEvent } from "$sb/app_event.ts";
 import { Decoration, syntaxTree } from "../deps.ts";
 import { Client } from "../client.ts";
@@ -8,6 +7,7 @@ import {
   invisibleDecoration,
   isCursorInRange,
 } from "./util.ts";
+import { commandLinkRegex } from "../../common/command.ts";
 
 /**
  * Plugin to hide path prefix when the cursor is not inside.